requirements.txt
pandas==2.2.2
numpy==1.26.4
pandas-ta==0.3.14b0
httpx==0.27.0
PyYAML==6.0.2
websockets==12.0
python-dateutil==2.9.0.post0
pytz==2024.1
loguru==0.7.2
